Duals test - prep work almost done

+++ src/org/interworldtransport/cladosG/Monad.java
@@ -1469,7 +1469,7 @@ public class Monad implements Modal {
 			if (getAlgebra().getGBasis().bladeOfGradeStream((byte) grade)
 					.filter(blade -> getScales().isNotZeroAt(blade)).parallel().findAny().isPresent()) {
 				foundGrades++;
-				gradeKey += Math.pow(10, grade);
+				gradeKey += (long) Math.pow(10, grade);
 			}
 		});
